refactor: move tests to correct sub-package
